Semin Branch + MinHwan Branch

+++ ChamEdu_v1/src/main/java/com/example/chamedu_v1/data/dto/ReviewMyPageResponseDto.java
@@ -9,7 +9,7 @@ public class ReviewMyPageResponseDto {
 
     private String title;
 
-    private float point;
+    private int point;
 
     private String content;
 }
